delete album options added with sweetalert2

+++ kenjoChallenge/src/app/pages/album/album.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ApiService } from 'src/app/shared/api.service';
 import { Album } from 'src/app/models/album';
+import Swal from 'sweetalert2/dist/sweetalert2.all.js';
 
 
 @Component({
@@ -36,11 +37,29 @@ export class AlbumComponent {
       });
   }
 
-  deleteAlbumById() {
-    this.apiService.deleteAlbum(this.ide).subscribe((data: Album) => {
-      console.log(data);
-      this.router.navigateByUrl('/albums');
-    });
-  }
+  
 
+  deleteAlbumById() {
+    Swal.fire({
+      title: '¿Seguro que quieres eliminar este album?',
+      text: "Asegúrate de eliminar el album que deseas",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: 'black',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+      
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Album eliminado!',
+          'El album ha sido eliminado correctamente...',
+          'success'
+        )
+        this.apiService.deleteAlbum(this.ide).subscribe((data: Album) => {
+          this.router.navigateByUrl('/albums');
+        });
+      }
+    });  
+  }
 }
